an alpha sufficient for internal usage

+++ src/csmalgos.hpp
@@ -8,7 +8,7 @@
 namespace lapis {
 
 	using csm_t = coord_t;
-	using taoid_t = double; //int64 would be ideal but none of the common raster formats support it
+	using taoid_t = uint32_t; //int64 would be ideal but none of the common raster formats support it
 
 	struct TaoIdMap {
 		struct XY { coord_t x, y; };
